server/requirements.txt
fastapi == 0.115.6
psycopg2-binary == 2.9.10
python-dotenv == 1.0.1
sqlalchemy == 2.0.37
uvicorn == 0.34.0
supabase == 2.11.0
pydantic == 2.10.5
pyjwt == 2.10.1
passlib == 1.7.4
google-auth == 2.37.0
httpx == 0.27.2
PyPDF2 == 3.0.1
python-multipart == 0.0.20
google-api-python-client == 2.159.0
google-auth-httplib2 == 0.2.0
google-auth-oauthlib == 1.2.1
requests == 2.32.3
python-jose == 3.3.0